corrected warnings about possible loss of precision for a conversion from double to float
svn path=/trunk/yarp2/; revision=8982

+++ src/yarpmanager/libymanager/src/kbase.cpp
@@ -1049,9 +1049,9 @@ float KnowledgeBase::calculateLoad(Computer* comp)
     if(!siblings) siblings = 1;
     //cout<<comp->getName()<<": "<<load.loadAverage1<<", "<<load.loadAverage5<<", "<<load.loadAverage15;
     //cout<<" ("<<siblings<<")"<<endl;    
-    float lavg = (load.loadAverage1*15.0 + 
-                  load.loadAverage5*10.0 + 
-                  load.loadAverage15*1.0) / 26.0;    
+    float lavg = (float)((load.loadAverage1*15.0 + 
+                          load.loadAverage5*10.0 + 
+                          load.loadAverage15*1.0) / 26.0);    
     return (lavg/(float)siblings);
 }
 
@@ -1268,10 +1268,10 @@ bool KnowledgeBase::reason(Graph* graph, Node* initial,
             // with load balancer tunning value;
             // It will help for better load distribution among nodes 
             Computer* comp = dynamic_cast<Computer*>(provider);
-            float default_tunning = 0.1;
+            float default_tunning = 0.1F;
             if(comp && (comp->getProcessor().getSiblings() > 0))
-                default_tunning = 1.0 / (float)comp->getProcessor().getSiblings();
-            float tunner = (module->getRank()<10)? default_tunning : (float)module->getRank()/100.0;
+                default_tunning = 1.0F / (float)comp->getProcessor().getSiblings();
+            float tunner = (module->getRank()<10)? default_tunning : (float)module->getRank()/100.0F;
             updateResourceWeight(*graph, provider, candidateLink->weight()+tunner);
         }
         resources.push_back(resource);
